pyproject.toml
[tool.poetry]
name = "musings"
version = "0.2.0"
description = "Compilation of my various projects for a portfolio"
authors = ["iTheSailor <112807438+iTheSailor@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.7.2"
beautifulsoup4 = "4.12.3"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
django = "4.2.9"
django-bootstrap-v5 = "1.0.11"
django-htmx = "1.17.2"
django-template-partials = "23.4"
django-widget-tweaks = "1.5.0"
fontawesomefree = "6.5.1"
googlemaps = "4.10.0"
idna = "3.6"
numpy = "1.26.3"
python-dotenv = "1.0.0"
requests = "2.31.0"
soupsieve = "2.5"
sqlparse = "0.4.4"
urllib3 = "2.1.0"
openmeteo-requests = "^1.2.0"
requests-cache = "^1.1.1"
retry-requests = "^2.0.0"
pandas = "^2.2.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^21.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
icecream = "^2.1.3"
aoe2netwrapper = "^0.3.1"
pillow = "^10.3.0"
yfinance = "^0.2.40"
psycopg2 = "^2.9.9"
uvicorn = "^0.30.1"
poetry-plugin-export = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
